HRRNHA algo added using Mean heuristic

+++ include/Process.h
@@ -14,7 +14,7 @@ public:
     int p_id;
     int completionTime;
     int responseTime;
-    int responseRatio;
+    float responseRatio;
 
 public:
     Process(int p_id, int arrivalTime, int burstTime);
